third_party/blink/web_tests/fast/forms/select-popup/popup-menu-resize-after-open-expected.txt

PASS popupWindow.innerWidth is 100
PASS popupWindow.innerWidth is 150
PASS successfullyParsed is true

TEST COMPLETE

o1
o2
